Update to move plugingl to libs
