+++ libs/plugingl/src/TexFont.cpp
