+++ libs/plugingl/src/qtstylesheet.cpp
